refactor: reorganize into more subpackages
