+++ app/setup_gcs_models.py
@@ -1,5 +1,6 @@
 import logging
 import os
+from collections import defaultdict
 
 from app import parser, utils, BASE
 from app.generator import trainer
@@ -13,7 +14,7 @@ def setup():
     Existing models will be overwritten.
     """
     logger.info("Downloading source files... ")
-    source_data_list = utils.download_all_source_files()
+    source_data_list = utils.common.download_all_source_files()
 
     logger.info("Creating description model...")
     description_text = " ".join([item["detailed_description"] for item in source_data_list])
@@ -37,7 +38,7 @@ def setup():
 
     # Train a dedicated model for each system requirement category
     logger.info("Creating system requirement models:")
-    extracted_requirement_map = utils.merge_requirements(source_data_list)
+    extracted_requirement_map = _merge_requirements(source_data_list)
     for key in extracted_requirement_map:
         logger.info(" # %s:", key)
         text_data = " ".join(extracted_requirement_map[key])
@@ -54,3 +55,39 @@ def setup():
     t.run()
 
     logger.info("Models saved in gs://%s", utils.MODEL_BUCKET)
+
+def _merge_requirements(source_data_list):
+    """Merge a list of requirement dicts.
+    Args:
+        source_data_list(list): list of requirements parsed from list of source description files
+    Return:
+        single dict with list of values from selected keys
+    """
+    requirements = ( [item["requirements"] for item in source_data_list if "requirements" in item] )
+
+    # Merge individual items in requirements
+    CATEGORIES_TO_EXTRACT = (
+        "OS",
+        "Processor",
+        "Memory",
+        "Graphics",
+        "DirectX",
+        "Storage",
+        "Hard Drive",
+        "Hard Disk Space",
+        "Sound Card",
+        "Additional",
+        "Additional Notes"
+    )
+
+    extracted_requirement_map = defaultdict(list)
+    for d in requirements:
+        for key, value in d.items():
+            if key in CATEGORIES_TO_EXTRACT:
+                extracted_requirement_map[key].extend(value)
+
+    # Merge keys denoting the same category
+    extracted_requirement_map["Storage"].extend(extracted_requirement_map.pop("Hard Drive", []))
+    extracted_requirement_map["Storage"].extend(extracted_requirement_map.pop("Hard Disk Space", []))
+    extracted_requirement_map["Additional Notes"].extend(extracted_requirement_map.pop("Additional", []))
+    return extracted_requirement_map
